Translate polymorphic maps to uninterpreted heaps (#60)
Detect types that match the structure of references, fields, and heaps,
and translate these into one monomorphic instance for each field type
that occurs in the program.

Along the way, this also:
* Makes sure block label names are escaped
* Avoids printing certain `free ensures` clauses with complex
quantifiers
* Adds missing support for translating bit vector types

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.