Don't drop axioms after analyzing procedures (#121)
Previously, there was a `pathConditions.pop` call at the end of the
`Procedure.eval` function that wound up removing axioms (which are added
to the path condition before analyzing procedures). Now there's a
corresponding `pathConditions.push` to ensure that the `pop` has
something to pop. Axioms are also now all added to the same scope with
`insert` rather than putting each in a new scope with `push`.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Shilpi Goel <shigoel@gmail.com>