Add a `distinct` command to Strata.Boogie and fix a few small issues (#166)
The new `distinct` command produces a special kind of axiom that is
preserved as a compact distinctness predicate throughout the
verification process. It may be possible to replace this with an
operator that can be used in normal `axiom` commands, but that would
require polymorphic lists in Strata.Lambda. So, for the moment, this
provides the functionality we need.

Along the way, this fixes fix a couple of Boogie issues:
* Sanitize names more consistently.
* Require spaces after command names for all top-level commands.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.