Normalize line endings in text files (#170)
We had a couple of files in the repo with different line endings. This
PR configures Git to automatically normalize line endings, and
normalizes the two files that were outliers.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.