Refactor type environment and unique identifier checks (#184)
*Description of changes:*

This PR makes 3 main changes:
1. Adds identifiers to the type environment, providing general-purpose
mechanisms (in `Identifiers.lean`) for checking dupliates more
efficiently (with a hash table rather than a list)
2. Splits the typing environment into 2 parts: `LContext`, which
provides contextual information that does not change during Lambda
typechecking (`Factory`, `KnownTypes`, and identifiers), and `TEnv`,
which does (inferred types, substitutions, etc).
3. Uses these changes to improve the typechecking for Boogie programs in
`ProgramType.lean`. In particular, this no longer requires a separate
check for `nodups` in the Program decl list; rather, it adds each
identifier to the context as it is seen, automatically checking for
uniqueness. This has the effect of performing more well-formedness
checks in Lambda typechecking rather than in Boogie.

There are also two smaller changes:
4. Due to the above, proving that well-typed Boogie programs have unique
names is no longer trivial, and this proof is added in `ProgramWF.lean`.
5. This PR does some other refactoring of `TEnv`, splitting it into the
name-generation part and the rest. This will be needed for the bounded
ints and/or function precondition PR, but is not immediately needed yet.

The bulk of the changes are in `Identifiers.lean`, `LExprTypeEnv.lean`,
`ProgramType.lean`, and `ProgramWF.lean`.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Josh Cohen <cohenjo@amazon.com>
Co-authored-by: Shilpi Goel <shigoel@gmail.com>